# api/openapi.yaml
openapi: 3.1.0
...
info:
    title: Nexus Open API
    description: |
        REST API for Nexus Open - an open-source iCUE Nexus companion app.
        Provides endpoints for device control, configuration management, and zone/plugin interactions.
    contact:
        name: Team
        url: https://github.com/mantonx/nexus-open
    version: 2.0.0
paths:
...
    /api/layout/draft:
        get:
            tags:
                - Layout
            summary: Get or open draft layout
            description: Returns the current draft, opening a new one from the committed config if none is active.
            operationId: getDraft
            responses:
                "200":
                    description: Draft layout
                    content:
                        application/json:
                            schema:
                                properties:
                                    active:
                                        type: boolean
                                    layout:
                                        type: object
                "503":
                    description: Draft manager not available
        put:
            tags:
                - Layout
            summary: Replace draft layout
            description: Replaces the entire draft with the provided config and previews it live on device.
            operationId: putDraft
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            description: Full zone.Config
            responses:
                "200":
                    description: Draft updated
                "422":
                    description: Invalid layout
    /api/layout/draft/zones:
        post:
            tags:
                - Layout
            summary: Add zone to draft
            description: Appends a zone to the specified page in the draft and redistributes widths.
            operationId: addDraftZone
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                page_index:
                                    type: integer
                                plugin:
                                    type: string
                                refresh_ms:
                                    type: integer
            responses:
                "200":
                    description: Zone added
                    content:
                        application/json:
                            schema:
                                properties:
                                    data:
                                        properties:
                                            id:
                                                type: string
                "422":
                    description: Zone cap exceeded or width redistribution impossible
...
servers:
    - url: http://localhost:1985
...